requirements.txt
flask
gunicorn
openai==0.28.1
jsonify
google-cloud-texttospeech
flask_cors